Containerize `Input` and `List`
Containerize `Input` and `List` to separate concerns.
