+++ src/InputContainer.jsx
@@ -1,0 +1,36 @@
+import React from 'react';
+
+import { useDispatch, useSelector } from 'react-redux';
+
+import Input from './Input';
+
+import { changeTaskTitle, addTask } from './actions';
+
+function selector(state) {
+  return {
+    taskTitle: state.taskTitle,
+    tasks: state.tasks,
+  };
+}
+
+export default function Page() {
+  const { taskTitle } = useSelector(selector);
+
+  const dispatch = useDispatch();
+
+  function handleChangeTitle(event) {
+    dispatch(changeTaskTitle(event.target.value));
+  }
+
+  function handleClickAddTask() {
+    dispatch(addTask());
+  }
+
+  return (
+    <Input
+      value={taskTitle}
+      onChange={handleChangeTitle}
+      onClick={handleClickAddTask}
+    />
+  );
+}
